[StudentNameVisibilityToggle] Move to components/material
This components is already written using MUI so it should go under /material
