+++ services/frontend/src/components/PopulationStudents/StudentTable/ProgressTab/index.jsx
@@ -7,8 +7,8 @@ import { Icon, Message, Tab } from 'semantic-ui-react'
 import { StudentInfoItem } from '@/components/common/StudentInfoItem'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import './index.css'
+import { useStudentNameVisibility } from '@/components/material/StudentNameVisibilityToggle'
 import { SortableTable } from '@/components/SortableTable'
-import { useStudentNameVisibility } from '@/components/StudentNameVisibilityToggle'
 import { ISO_DATE_FORMAT } from '@/constants/date'
 import { useGetSemestersQuery } from '@/redux/semesters'
