+++ services/frontend/src/components/material/StudentNameVisibilityToggle.tsx
