+++ services/frontend/src/pages/Students/StudentDetails/StudentInfoCard.tsx
@@ -3,7 +3,7 @@ import { Box, Card, CardContent, Typography, Button, Stack } from '@mui/material
 
 import { callApi } from '@/apiConnection'
 import { ExternalLink } from '@/components/material/Footer/ExternalLink'
-import { useStudentNameVisibility } from '@/components/StudentNameVisibilityToggle'
+import { useStudentNameVisibility } from '@/components/material/StudentNameVisibilityToggle'
 import { sisUrl } from '@/conf'
 import { DISPLAY_DATE_FORMAT, DISPLAY_DATE_FORMAT_DEV } from '@/constants/date'
 import { useGetAuthorizedUserQuery } from '@/redux/auth'
