+++ services/frontend/src/pages/Students/StudentSearch.tsx
@@ -16,8 +16,8 @@ import { useEffect, useMemo, useState } from 'react'
 import { useNavigate } from 'react-router'
 
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
+import { useStudentNameVisibility } from '@/components/material/StudentNameVisibilityToggle'
 import { StyledTable } from '@/components/material/StyledTable'
-import { useStudentNameVisibility } from '@/components/StudentNameVisibilityToggle'
 import { DISPLAY_DATE_FORMAT } from '@/constants/date'
 import { useSearchStudentsQuery } from '@/redux/students'
 import { reformatDate } from '@/util/timeAndDate'
